Moved requirements data model to new file
